fix an arithmetic overflow
git-svn-id: http://svn.eurecom.fr/openair4G/trunk@7872 818b1a75-f10b-46b9-bf7c-635c3b92a50f

+++ openair1/SCHED/phy_procedures_lte_eNb.c
@@ -3218,7 +3218,7 @@ void phy_procedures_eNB_RX(const unsigned char sched_subframe,PHY_VARS_eNB *phy_
   int16_t n1_pucch0,n1_pucch1,n1_pucch2,n1_pucch3;
   uint8_t do_SR = 0;
   uint8_t pucch_sel = 0;
-  int16_t metric0=0,metric1=0;
+  int32_t metric0=0,metric1=0;
   ANFBmode_t bundling_flag;
   PUCCH_FMT_t format;
   uint8_t nPRS;
